Pruning dependencies
