+++ test/quirk/add-emoji.js
@@ -1,5 +1,5 @@
 const expect = require("chai").expect;
-const Quirk = require("../../quirk");
+const Quirk = require("../../index");
 
 describe("add emoji", function () {
 /*     it("should add emoji to both plain and quirk substitution lists", function () {
