+++ test/quirk/add-strip-pattern.js
@@ -1,5 +1,5 @@
 const expect = require('chai').expect;
-const Quirk = require('../../quirk');
+const Quirk = require("../../index");
 
 describe('add plain strip pattern', function () {
