+++ test/quirk/set-sentence-case-upper.js
@@ -1,5 +1,5 @@
 const expect = require("chai").expect;
-const Quirk = require("../../quirk");
+const Quirk = require("../../index");
 
 describe("setSentenceCase - uppercase", function () {
     it('should enforce uppercase', function () {
