+++ test/quirk/set-separator.js
@@ -1,5 +1,5 @@
 const expect = require('chai').expect;
-const Quirk = require('../../quirk');
+const Quirk = require("../../index");
 
 describe('separator', function () {
     it('should create a valid quirk from a separator', function () {
