+++ test/use-cases/contractions-experimental.js
@@ -1,5 +1,5 @@
 const expect = require("chai").expect;
-const Quirk = require("../../quirk");
+const Quirk = require("../../index");
 
 const contractionList = [
     "ain't",
